Extract new build_context module out of context module
